cleanup, early purge tower stop if no longer necessary
